modified Makefiles to install include files
Summary:
The current Makefiles do not install the SLS headers into the system, forcing us to point to the source directory when building external utilities. The Makefile for tests also points to a nonexistent directory, stopping the install process until the user manually
creates it, even though we do not need to install them in the first place. Fix the Makefiles to install the include files and install the tests to /tmp from where we can easilydelete them.

Test Plan: None, nonfunctionali change

Reviewers: ali, ryan, tavianator

Differential Revision: https://review.rcs.uwaterloo.ca/D770