fix crash caused by faulty SLS kldload error handling
Summary:
When unloading, the SLS traverses a list that holds all processes in Aurora and kills them.
This list is initialized when the module is loaded, but there are cases when the load fails before the
list is initialized. In that case the unload routine attempts to traverse the uninitialized list and
crashes the system. Add a guard in the code to prevent the list from being accessed if uninitialized.

Test Plan: XXX Add a testcase that triggers this to ensure we stamped it out

Reviewers: ali, ryan, tavianator

Differential Revision: https://review.rcs.uwaterloo.ca/D778