Object Awareness to Allocator
To be able to properly move data for a given thread, its required
that it knows the object, as well as offet that exists (i.e., a reverse
mapping). This can be
embedded into the allocator state itself. This only adds about 16 bytes
of overahead per 4KiB block (size and offset are already encoded in the
allocator).